Crud.Domain/Handlers/CreatePersonCommandHandler.cs
using Crud.Domain.Commands;
using Crud.Domain.Models;
using Crud.Repository;
using MediatR;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Crud.Domain.Handlers
{
    public class CreatePersonCommandHandler : IRequestHandler<CreatePersonCommand, int>
    {
        private readonly ApplicationDbContext _context;

        public CreatePersonCommandHandler(ApplicationDbContext context)
        {
            _context = context;
        }

        public async Task<int> Handle(CreatePersonCommand request, CancellationToken cancellationToken)
        {
            var entity = new Person { Name = request.Name, Age = request.Age };
            _context.Persons.Add(entity);
            await _context.SaveChangesAsync(cancellationToken);
            return entity.Id;
        }
    }
}
